Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main()
 {
-    int a;
-    int b;
-    int c;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    double a;
+    double b;
+    double c;
+    scanf("%lf",&a);
+    scanf("%lf",&b);
+    scanf("%lf",&c);
     float avg=(a+b+c)/3;
     printf("Average: %.2lf",avg);
